// code/frontend/CNet.Web.Admin/src/api/admin/gen/data-contracts.ts
export interface PubFunction {
  functionCode?: string | null
  functionChina?: string | null
  functionDescrip?: string | null
  parentCode?: string | null
  menuFlag?: boolean
  stopFlag?: boolean | null
  urlString?: string | null
  /** @format date-time */
  lmdt?: string | null
  lmid?: string | null
  /** @format int32 */
  sortId?: number | null
  pageTarget?: string | null
  menuIcon?: string | null
  routerPath?: string | null
  isCache?: boolean | null
}
// ...
export type ResCode = 0 | 1 | -1
// ...
export interface BooleanDataRes {
  code?: ResCode
  msg?: string | null
  data?: boolean
}

export interface PubDepartmentDataRes {
  code?: ResCode
  msg?: string | null
  data?: PubDepartment
}

export interface PubFunctionListDataRes {
  code?: ResCode
  msg?: string | null
  data?: PubFunction[] | null
}

export interface VPubfunctionParent {
  functionCode?: string | null
  functionChina?: string | null
  functionDescrip?: string | null
  parentCode?: string | null
  menuFlag?: boolean
  stopFlag?: boolean | null
  urlString?: string | null
  /** @format date-time */
  lmdt?: string | null
  lmid?: string | null
  /** @format int32 */
  sortId?: number | null
  pageTarget?: string | null
  menuIcon?: string | null
  routerPath?: string | null
  isCache?: boolean | null
  parentName?: string | null
}

export interface VPubfunctionParentListDataRes {
  code?: ResCode
  msg?: string | null
  data?: VPubfunctionParent[] | null
}

export interface Meta {
  /** 图标 */
  icon?: string | null
  /** 不缓存 */
  notCache?: boolean | null
  /** 标题 */
  title?: string | null
  /** 权限列表 */
  access?: string[] | null
  /** 是否新窗口 */
  isNewWindow?: boolean
}

export interface Menu {
  /** 路由路径 */
  path?: string | null
  /** 名称 */
  name?: string | null
  meta?: Meta
  /** 子组件 */
  children?: Menu[] | null
  component?: string | null
}

export interface MenuListDataRes {
  code?: ResCode
  msg?: string | null
  data?: Menu[] | null
}

export interface PubFunctionDataRes {
  code?: ResCode
  msg?: string | null
  data?: PubFunction
}